implement api WITHOUT relations table

+++ api/src/main/java/fr/video_game_tournament/api/repositories/VideoGameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VideoGameRepository extends CrudRepository<VideoGame, Long> {
+public interface VideoGameRepository extends CrudRepository<VideoGame, Integer> {
 }
